refactor: re-organized controllers
