+++ src/main/java/it/gov/pagopa/fdr/controller/PspsController.java
@@ -1,4 +1,4 @@
-package it.gov.pagopa.fdr.controller.psps;
+package it.gov.pagopa.fdr.controller;
 
 import it.gov.pagopa.fdr.Config;
 import it.gov.pagopa.fdr.controller.model.common.response.GenericResponse;
@@ -21,16 +21,7 @@ import jakarta.validation.Valid;
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Pattern;
-import jakarta.ws.rs.Consumes;
-import jakarta.ws.rs.DELETE;
-import jakarta.ws.rs.DefaultValue;
-import jakarta.ws.rs.GET;
-import jakarta.ws.rs.POST;
-import jakarta.ws.rs.PUT;
-import jakarta.ws.rs.Path;
-import jakarta.ws.rs.PathParam;
-import jakarta.ws.rs.Produces;
-import jakarta.ws.rs.QueryParam;
+import jakarta.ws.rs.*;
 import jakarta.ws.rs.core.MediaType;
 import java.time.Instant;
 import org.eclipse.microprofile.openapi.annotations.Operation;
@@ -43,13 +34,13 @@ import org.eclipse.microprofile.openapi.annotations.tags.Tag;
 import org.jboss.logging.Logger;
 import org.jboss.resteasy.reactive.RestResponse;
 
-@Tag(name = "Internal PSP", description = "PSP operations")
-@Path("/internal/psps/{" + AppConstant.PSP + "}")
+@Tag(name = "PSP", description = "PSP operations")
+@Path("/psps/{" + AppConstant.PSP + "}")
 @Consumes("application/json")
 @Produces("application/json")
-public class InternalPspsResource extends BasePspResource {
+public class PspsController extends BasePspController {
 
-  protected InternalPspsResource(
+  protected PspsController(
       Logger log,
       Config config,
       PspsValidationService validator,
@@ -59,7 +50,7 @@ public class InternalPspsResource extends BasePspResource {
     super(log, config, validator, internalValidator, mapper, service);
   }
 
-  @Operation(operationId = "internalCreate", summary = "Create fdr", description = "Create fdr")
+  @Operation(operationId = "create", summary = "Create fdr", description = "Create fdr")
   @RequestBody(content = @Content(schema = @Schema(implementation = CreateFlowRequest.class)))
   @APIResponses(
       value = {
@@ -76,8 +67,8 @@ public class InternalPspsResource extends BasePspResource {
       })
   @POST
   @Path("/fdrs/{" + AppConstant.FDR + "}")
-  @Re(action = FdrActionEnum.INTERNAL_CREATE_FLOW)
-  public RestResponse<GenericResponse> internalCreate(
+  @Re(action = FdrActionEnum.CREATE_FLOW)
+  public RestResponse<GenericResponse> create(
       @PathParam(AppConstant.PSP) String pspId,
       @PathParam(AppConstant.FDR) @Pattern(regexp = "[a-zA-Z0-9\\-_]{1,35}") String fdr,
       @NotNull @Valid CreateFlowRequest createRequest) {
@@ -86,7 +77,7 @@ public class InternalPspsResource extends BasePspResource {
   }
 
   @Operation(
-      operationId = "internalAddPayment",
+      operationId = "addPayment",
       summary = "Add payments to fdr",
       description = "Add payments to fdr")
   @RequestBody(content = @Content(schema = @Schema(implementation = AddPaymentRequest.class)))
@@ -105,8 +96,8 @@ public class InternalPspsResource extends BasePspResource {
       })
   @PUT
   @Path("/fdrs/{" + AppConstant.FDR + "}/payments/add")
-  @Re(action = FdrActionEnum.INTERNAL_ADD_PAYMENT)
-  public GenericResponse internalAddPayment(
+  @Re(action = FdrActionEnum.ADD_PAYMENT)
+  public GenericResponse addPayment(
       @PathParam(AppConstant.PSP) String pspId,
       @PathParam(AppConstant.FDR) String fdr,
       @NotNull @Valid AddPaymentRequest addPaymentRequest) {
@@ -114,7 +105,7 @@ public class InternalPspsResource extends BasePspResource {
   }
 
   @Operation(
-      operationId = "internalDeletePayment",
+      operationId = "deletePayment",
       summary = "Delete payments to fdr",
       description = "Delete payments to fdr")
   @RequestBody(content = @Content(schema = @Schema(implementation = DeletePaymentRequest.class)))
@@ -133,15 +124,15 @@ public class InternalPspsResource extends BasePspResource {
       })
   @PUT
   @Path("/fdrs/{" + AppConstant.FDR + "}/payments/del")
-  @Re(action = FdrActionEnum.INTERNAL_DELETE_PAYMENT)
-  public GenericResponse internalDeletePayment(
+  @Re(action = FdrActionEnum.DELETE_PAYMENT)
+  public GenericResponse deletePayment(
       @PathParam(AppConstant.PSP) String pspId,
       @PathParam(AppConstant.FDR) String fdr,
       @NotNull @Valid DeletePaymentRequest deletePaymentRequest) {
     return baseDeletePayment(pspId, fdr, deletePaymentRequest);
   }
 
-  @Operation(operationId = "internalPublish", summary = "Publish fdr", description = "Publish fdr")
+  @Operation(operationId = "publish", summary = "Publish fdr", description = "Publish fdr")
   @APIResponses(
       value = {
         @APIResponse(ref = "#/components/responses/InternalServerError"),
@@ -157,13 +148,13 @@ public class InternalPspsResource extends BasePspResource {
       })
   @POST
   @Path("/fdrs/{" + AppConstant.FDR + "}/publish")
-  @Re(action = FdrActionEnum.INTERNAL_PUBLISH)
-  public GenericResponse internalPublish(
+  @Re(action = FdrActionEnum.PUBLISH)
+  public GenericResponse publish(
       @PathParam(AppConstant.PSP) String pspId, @PathParam(AppConstant.FDR) String fdr) {
-    return basePublish(pspId, fdr, true);
+    return basePublish(pspId, fdr, false);
   }
 
-  @Operation(operationId = "internalDelete", summary = "Delete fdr", description = "Delete fdr")
+  @Operation(operationId = "delete", summary = "Delete fdr", description = "Delete fdr")
   @APIResponses(
       value = {
         @APIResponse(ref = "#/components/responses/InternalServerError"),
@@ -179,16 +170,16 @@ public class InternalPspsResource extends BasePspResource {
       })
   @DELETE
   @Path("/fdrs/{" + AppConstant.FDR + "}")
-  @Re(action = FdrActionEnum.INTERNAL_DELETE_FLOW)
-  public GenericResponse internalDelete(
+  @Re(action = FdrActionEnum.DELETE_FLOW)
+  public GenericResponse delete(
       @PathParam(AppConstant.PSP) String pspId, @PathParam(AppConstant.FDR) String fdr) {
     return baseDelete(pspId, fdr);
   }
 
   @Operation(
-      operationId = "internalGetAllCreated",
-      summary = "Get all fdr inserted",
-      description = "Get all fdr inserted")
+      operationId = "getAllcreated",
+      summary = "Get all fdr created",
+      description = "Get all fdr created")
   @APIResponses(
       value = {
         @APIResponse(ref = "#/components/responses/InternalServerError"),
@@ -204,8 +195,8 @@ public class InternalPspsResource extends BasePspResource {
       })
   @GET
   @Path("/created")
-  @Re(action = FdrActionEnum.INTERNAL_GET_ALL_CREATED_FDR)
-  public PaginatedFlowsCreatedResponse internalGetAllCreated(
+  @Re(action = FdrActionEnum.GET_ALL_CREATED_FDR)
+  public PaginatedFlowsCreatedResponse getAllCreated(
       @PathParam(AppConstant.PSP) String pspId,
       @QueryParam(AppConstant.CREATED_GREATER_THAN) Instant createdGt,
       @QueryParam(AppConstant.PAGE) @DefaultValue(AppConstant.PAGE_DEAFULT) @Min(value = 1)
@@ -216,7 +207,7 @@ public class InternalPspsResource extends BasePspResource {
   }
 
   @Operation(
-      operationId = "internalGetCreated",
+      operationId = "getCreated",
       summary = "Get created fdr",
       description = "Get created fdr")
   @APIResponses(
@@ -234,8 +225,8 @@ public class InternalPspsResource extends BasePspResource {
       })
   @GET
   @Path("/created/fdrs/{" + AppConstant.FDR + "}/organizations/{" + AppConstant.ORGANIZATION + "}")
-  @Re(action = FdrActionEnum.INTERNAL_GET_CREATED_FDR)
-  public SingleFlowCreatedResponse internalGetCreated(
+  @Re(action = FdrActionEnum.GET_CREATED_FDR)
+  public SingleFlowCreatedResponse getCreated(
       @PathParam(AppConstant.PSP) String psp,
       @PathParam(AppConstant.FDR) String fdr,
       @PathParam(AppConstant.ORGANIZATION) String organizationId) {
@@ -243,9 +234,9 @@ public class InternalPspsResource extends BasePspResource {
   }
 
   @Operation(
-      operationId = "internalGetCreatedPayment",
-      summary = "Get internal created payments of fdr",
-      description = "Get internal created payments of fdr")
+      operationId = "getCreatedPayment",
+      summary = "Get created payments of fdr",
+      description = "Get created payments of fdr")
   @APIResponses(
       value = {
         @APIResponse(ref = "#/components/responses/InternalServerError"),
@@ -266,8 +257,8 @@ public class InternalPspsResource extends BasePspResource {
           + "}/organizations/{"
           + AppConstant.ORGANIZATION
           + "}/payments")
-  @Re(action = FdrActionEnum.INTERNAL_GET_CREATED_FDR_PAYMENT)
-  public PaginatedPaymentsResponse internalGetCreatedPayment(
+  @Re(action = FdrActionEnum.GET_CREATED_FDR_PAYMENT)
+  public PaginatedPaymentsResponse getCreatedPayment(
       @PathParam(AppConstant.PSP) String psp,
       @PathParam(AppConstant.FDR) String fdr,
       @PathParam(AppConstant.ORGANIZATION) String organizationId,
@@ -279,9 +270,9 @@ public class InternalPspsResource extends BasePspResource {
   }
 
   @Operation(
-      operationId = "internalGetAllPublishedByPsp",
-      summary = "Get all internal fdr published",
-      description = "Get all internal fdr published")
+      operationId = "getAllPublishedByPsp",
+      summary = "Get all fdr published",
+      description = "Get all fdr published")
   @APIResponses(
       value = {
         @APIResponse(ref = "#/components/responses/InternalServerError"),
@@ -297,8 +288,8 @@ public class InternalPspsResource extends BasePspResource {
       })
   @GET
   @Path("/published")
-  @Re(action = FdrActionEnum.INTERNAL_GET_ALL_FDR_PUBLISHED_BY_PSP)
-  public PaginatedFlowsPublishedResponse internalGetAllPublishedByPsp(
+  @Re(action = FdrActionEnum.GET_ALL_FDR_PUBLISHED_BY_PSP)
+  public PaginatedFlowsPublishedResponse getAllPublishedByPsp(
       @PathParam(AppConstant.PSP) String idPsp,
       @QueryParam(AppConstant.ORGANIZATION) @Pattern(regexp = "^(.{1,35})$") String organizationId,
       @QueryParam(AppConstant.PUBLISHED_GREATER_THAN) Instant publishedGt,
@@ -310,9 +301,9 @@ public class InternalPspsResource extends BasePspResource {
   }
 
   @Operation(
-      operationId = "internalGetPublishedByPsp",
-      summary = "Get internal fdr Published",
-      description = "Get internal fdr Published")
+      operationId = "getPublishedByPsp",
+      summary = "Get fdr Published",
+      description = "Get fdr Published")
   @APIResponses(
       value = {
         @APIResponse(ref = "#/components/responses/InternalServerError"),
@@ -335,19 +326,19 @@ public class InternalPspsResource extends BasePspResource {
           + "}/organizations/{"
           + AppConstant.ORGANIZATION
           + "}")
-  @Re(action = FdrActionEnum.INTERNAL_GET_FDR_PUBLISHED_BY_PSP)
-  public SingleFlowResponse internalGetPublishedByPsp(
+  @Re(action = FdrActionEnum.GET_FDR_PUBLISHED_BY_PSP)
+  public SingleFlowResponse getPublishedByPsp(
       @PathParam(AppConstant.PSP) String psp,
       @PathParam(AppConstant.FDR) String fdr,
       @PathParam(AppConstant.REVISION) Long rev,
       @PathParam(AppConstant.ORGANIZATION) String organizationId) {
-    return baseGetPublished(organizationId, fdr, rev, psp);
+    return baseGetPublished(psp, fdr, rev, organizationId);
   }
 
   @Operation(
-      operationId = "internalGetPaymentPublishedByPSp",
-      summary = "Get internal payments of fdr Published",
-      description = "Get internal payments of fdr Published")
+      operationId = "getPaymentPublishedByPsp",
+      summary = "Get payments of fdr Published",
+      description = "Get payments of fdr Published")
   @APIResponses(
       value = {
         @APIResponse(ref = "#/components/responses/InternalServerError"),
@@ -370,8 +361,8 @@ public class InternalPspsResource extends BasePspResource {
           + "}/organizations/{"
           + AppConstant.ORGANIZATION
           + "}/payments")
-  @Re(action = FdrActionEnum.INTERNAL_GET_FDR_PAYMENT_PUBLISHED_BY_PSP)
-  public PaginatedPaymentsResponse internalGetPaymentPublishedByPSp(
+  @Re(action = FdrActionEnum.GET_FDR_PAYMENT_PUBLISHED_BY_PSP)
+  public PaginatedPaymentsResponse getPaymentPublishedByPsp(
       @PathParam(AppConstant.PSP) String psp,
       @PathParam(AppConstant.FDR) String fdr,
       @PathParam(AppConstant.REVISION) Long rev,
@@ -380,6 +371,6 @@ public class InternalPspsResource extends BasePspResource {
           long pageNumber,
       @QueryParam(AppConstant.SIZE) @DefaultValue(AppConstant.SIZE_DEFAULT) @Min(value = 1)
           long pageSize) {
-    return baseGetFdrPaymentPublished(organizationId, fdr, rev, psp, pageNumber, pageSize);
+    return baseGetFdrPaymentPublished(psp, fdr, rev, organizationId, pageNumber, pageSize);
   }
 }
