+++ src/test/java/it/gov/pagopa/fdr/controller/info/InfoControllerIT.java
@@ -3,6 +3,6 @@ package it.gov.pagopa.fdr.controller.info;
 import io.quarkus.test.junit.QuarkusIntegrationTest;
 
 @QuarkusIntegrationTest
-class InfoResourceIT extends InfoResourceTest {
+class InfoControllerIT extends InfoControllerTest {
   // Execute the same tests but in packaged mode.
 }
