+++ src/test/java/it/gov/pagopa/fdr/controller/info/InfoControllerTest.java
@@ -7,7 +7,7 @@ import io.quarkus.test.junit.QuarkusTest;
 import org.junit.jupiter.api.Test;
 
 @QuarkusTest
-class InfoResourceTest {
+class InfoControllerTest {
 
   String name = "pagopafdr";
